Extract constants into separate modules
