+++ daylatro.cabal
@@ -40,6 +40,7 @@ common library
 
 common executable
   import: library
+  build-depends: daylatro
   ghc-options:
     -rtsopts
     -threaded
@@ -67,6 +68,10 @@ library
   -- cabal-gild: discover source/library
   exposed-modules:
     Daylatro
+    Daylatro.Constant.Font
+    Daylatro.Constant.Script
+    Daylatro.Constant.Shader
+    Daylatro.Constant.Style
 
   hs-source-dirs: source/library
   other-modules: PackageInfo_daylatro
@@ -74,23 +79,6 @@ library
 executable daylatro
   import: executable
   autogen-modules: PackageInfo_daylatro
-  build-depends:
-    bytestring ^>=0.12.2.0,
-    containers ^>=0.8,
-    data-default ^>=0.8.0.1,
-    exceptions ^>=0.10.9,
-    formatting ^>=7.2.0,
-    http-types ^>=0.12.4,
-    lucid2 ^>=0.0.20250303,
-    network-uri ^>=2.6.4.2,
-    random ^>=1.3.1,
-    sqlite-simple ^>=0.4.19.0,
-    text ^>=2.1.2,
-    time ^>=1.14,
-    wai ^>=3.2.4,
-    warp ^>=3.4.8,
-    xml-conduit ^>=1.10.0.1,
-
   hs-source-dirs: source/executable
   main-is: Main.hs
   other-modules: PackageInfo_daylatro
