+++ source/library/Daylatro/Constant/Font.hs
@@ -1,0 +1,9 @@
+module Daylatro.Constant.Font where
+
+import qualified Data.Text as Text
+
+-- | <https://fontstruct.com/fontstructions/show/2326420/balatro>
+balatro :: Text.Text
+balatro =
+  Text.pack
+    "d09GRk9UVE8AABLcAAsAAAAAJrAAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAABDRkYgAAAFkAAADF4AABrSSIGO3UZGVE0AABIQAAAAGgAAAByfBi+AR0RFRgAAEfAAAAAdAAAAHgAnAKdPUy8yAAABZAAAAEgAAABgdJZyRGNtYXAAAARUAAABJQAAAZLDHprtaGVhZAAAAQgAAAAzAAAANihLn+FoaGVhAAABPAAAAB8AAAAkDOgHm2htdHgAABIsAAAArgAAAoK3/EGubWF4cAAAAVwAAAAGAAAABgChUABuYW1lAAABrAAAAqUAAAXlE2CrhXBvc3QAAAV8AAAAEgAAACAAmQBkeJxjYGRgYABisY9O4fH8Nl8ZuJlfAEUYHgmvvguj/x79e5TdgbMWyOVgYAKJAgBoqQ2NAHicY2BkYGB/yAAE7I5/j/7fye7AwMiAChYAAHf0BZQAAABQAAChAAB4nGNgZglhnMDAysDC/IL5BQMQQGggTgGKAwELAxwwMyABt+CQIAYHBgUFWfaHID6IVGBk/H/2PwMDKyu7I4MCEDICAJqrDbV4nKWUS2sTURTH/5NJH0lbRSoFQctB3IiSSaYllOKuUFxUCy10oSBMppNkmmRuOvdO03ThskuXfhLxCwhuXKl7wZUgfgFdeebm2KYlRcSETH73vOa8ZgAsOwkcjD63sCrsYAYvhAvMR8Iu8yvhIq7jrfAU5vBFeJptfgnP4IHzTHgWS84b4RLKzlfhMu4V6sJzWC98FJ7HI/dUeAE33G/Ci7hffMKZOMUSn3ybVc4OruGpcIG5KewyvxQuYhmvhaewhPfC02zzXXgGe05BeBYPnVPhEm46H4TLeOz8EJ6DKTwXnsexOy+8gLvuO+FF7Lg/sQGFPoZIEaOFNgwIXZaFCPjfsCRDgg5LfVT5twJsqP4wjVttQ10VBl3TzpIO+VWfVQ3rFbBfyjHQCLqBSRl2EHH0zOpSPkatrBukV9tflNNlf7Gjszh7rE+huQbF2RJqqHC22ItSHauEapXq1ffatD4Gu1aTceV5D2KORmxNVhpgn+/Qs3fPe6F4kjTRs8JXxZbYVInZNWkWGoo1BWTSYD/qBWmHVJPOlZVQsXHed8Nz0FiHx9+mRNaTIreN6et1z2tyED0WZNLYcD4i0O2LPaA7XMOAT3mlDXaIrWsui22IyRWWgU/SvM80CDQ1srhraBCb9lhd5X8rarJ2NE7NWs3RFCfm2RX0UefXQr6Q3uWxTm7O2JH3QXu6rQaev+LXV/2qd7YJ/5+wslt4Od0ap1vDGrM3eUh/TVqlZznX6rU13xsbK7Y55ciu/Z9xEba4d6GVar5iux8ldji0FYdRoqN87WN+hUZ29PmjTzjkdAL26vA5sZIT1vc5zaEti3DA8Ucr0mPZaHmObYSIHxBsxkcRNeIWHWZB2ImTFp1E/fYw1XSgeEN6Q16XY2pE+/gNQGcqxQAAAHicY2BgYGaAYBkGRgYQ6AHyGMF8FoYCIC3BIAAU4WCoY7jG8JThPcM3hj8M/xnzGSsUJBVk//8HqlBgOMBwg+E5w0eGH0CZPKCMhILM////H/5f8H/+/3n/5/6f83/Wf53/Kg9aHzRDbcEKGNkY8ElD1DAxs7CysXNwcnHz8PLxCwgKCYuIiolLSEpJy8jKySsoKimrqKqpa2hqaevo6ukbGBoZm5iamVtYWlnb2NrZOzg6Obu4url7eHp5+/j6+QcEBgWHhIaFR0RGRcfExsUzpKSmZxaWVTfUNzY3tbS2t3V0dnf19PZN6J84edK0qdNnQJ1RB6HSytGcNwWJnZBcCqLmL5g7j4Fh5hwkmaSsxOyM3Lz8nOIShqLKqgokKQAaxlQpAAAAeJxjYGaAgGkMKQxYAAAWKwD+AAB4nK1Ya5AcVRU+PdvTvdsz27uzO72AdsKmJzwSCNk8IIEQmeEpr4AIAgqTigaWVFKESgVQHBRSWAJT8kOBovBRRZU/UrIK3SASIAKiFuHhrEBEeajxgT90SiypTM90z+I599zu6ZmdlFLlTu1MP+499zy+851zrwKqCoqiDH5+07ZNO3dsByUFCpzoLwB/oeIfmfInB/zD1cZV2YF/ZtWwFtZsvWgP3Zq24cERG2DUVt7O2bDAPvziMdBoqgYZMOEYOAnWX7Zjy/XTN9LXihOmzth+w5d2bJm+bufktu1f2LRt53U3Xr91cuXUylVy4clLrpm+cdumHZEe+KdACgZAhTQK1eE6GAIDhWdhGBcYgVHIwRiMQx4smIDD4HA4Aj4GHwfUBhbCkTAJi8CBAiyGo+BoVOhYWAJL4Tg4HpbBCbAcpmAFrIRVsBpORGXXwFo4GU6BdXAqrIdPwGlQhAqcDmfAmXAWnA3nwCfhXDgPzocL4ELYABfBxfApuAQ+DZfCZfAZuByugCvhs/A5uAquhjJshB/BwzADj8IP4RG4G1z4MXjwGDwOe+AJ+Ak8CQ/CU/BTeBqegWdhL3wDfg7Pw8/gBfgFfBfug5fhRdgHr8JL8Ar8Cl6DGszCr+E38Dq8AfvhTXgbfgu/g3fgLfgWHIDfwx/gj/Bn5evKnfBLKMEgbIGb4SGKx7mo8CbYjgs1ldOVHyiPKa8p76dyqanUGanzU5ekvpa6N/Vo6oOBqYFrB+4feGXgA3WFeoU6rd6u7lFfVOfSp6V3p/ek39Py2kptjbZBu037jvaUdkA/Wl+rb9Gf1l/S64NLB68cvG9wZtAbfG/o2KFbh7yhvw4FxnLjLGOHUTEeMHYb+4ya8b7xYcbJXJTZmrkl8+3M65kD2cHscNbOrs9em709O5OtZfcPK8Nrh+8Yvnt473BgGuY55mbzK+b95qvmv0bMkXUju0ZeGHlt5MBIODo6umz0vNGto7eNPp6zc6tzF+RuyH0z98hYaiw3Njm2buyisfLYnWNvjDXHF46fM75lvDr+xPje8VfH3x2vj7fzw/kF+WX59fkz85fmN+d35u/I35v/fv7hvJffk38uvy8/m38z//f8B/kPrQlruW/Y6Zf3Wu3LW8X25XPF9Be/ZzXFdauYttMztr/bqtr6Fcl3duuWavps27/YsvU/2dqo7S/UzfAmcy5XVSeNwPNLlp2+t+HZrW262d5npzV1kdEoTCwyzMY9zXqzbPmlwAtLeuCGNavhBV7DbXj0rEEft+EGrvj1Alen1zSYbgJxzU80M9jf8CZ4tl9reGGxUQgKeFUQH3waeDQb10CVigE+xKuC+ODsZnli0pir2On7LLlSgApvRXVwsisE0L8j72l1nKfTl/h3+IFmHmyWGx7KoAUCh3QIa3rDbdYtXJQswSckO7jLokscjXIjRf1arBcqi1+su5BN8sg6obJmzv2NHWm2d6Ev6amtSVNJJbsV0Hh0+K6qSkrb6XRAFngkyG6+S1KicDgUDscg4VaBfhxLPhJe9WVUZDwilwvp+2hlYQm9xCDijGZdXYx67aPwm1XVL/W4kK8xQodwofDgUQbdWEcjSNwJGt6qMDBIBdS/IdZPSOePI9xYRNkODdfptlkmm9tlGW3+4Gp+TSMZ0n3sG7+IMRBw6yuEQxMJI1VbFYIvSxSSdb84V7Ga5XbZn00qRd9+yfaf7bNiWBNYqJON7XJY1BkI7Xq0CjrkGHZIItKsEuVBwwlRW5qkd2ZhOjC6ItS2y+RFobonXR/WYul+ETF6LC7SH5pJ+/WkA2z1ZpzmF2lasy4tKXem2M0NOnJBoyDIALMdQUG3BBtPwmYWMwIdiW/CzX4JCaPhCNMK0k2FRHxdyk32tyeHuZT9fg1R1677s9ZiFqPGsPCkIMpbkZI0T+RXInCOZk77pQnJFIiOEmasDB1jqkauxmdkWp1dncQUajFNulpLDPICpZCLqRU4KmOecp9VQkZo10Vw2mWd86tAjCEkWuw6iqM+V8GYOIZfw1TE6FGiiSDHbmciWWr0BlE8tvUS6dmL2hh/fWGExEEmMtF20IfpiA86ImMGqarHUZLjiwi8qKEX1lD7/2EdCiVxfQQnIaiBYtAQXBazBSFTD4skLfx3F5PRMoFDPOdQFk93JRVSRDqNUfDQTaQ41p5IXbFAWJNochJcxOkbFBLQI5d5nIyMHI00CfZLMegFUkM3/Rl8onkxh3sRlyP6AxcdEJZwItuCkzTTXxBPEPpKIDPni3+iHaqbLolwOIARjgMxSRNwFmS/Ve9lzW6ExHFjR3N4pqlOWsIylyqUCIren2XQgcn604vAbvkMNS8KeAIK8ygvJgq9l8k6REZpKVwowxmW0IVC5PGGSSlLhR4D6M8KgR5lSjjbcIh/m5RgpFSILznKyIz3xMj0RF/CUPAJGCWONHpep/hu4AahEdVfwRaoUqfwEPuzUbLpIL4lc6I6J2ZhyEg0DQixs9Foll+MKhG91IWd7EX+J1vYTUwbfk3aEn3ada1vfSXgtyp96iuBVROJhAETCeRSAhHhUQfEmcOYiv7JrAiPTgecZIYWpaRHKelRSoryRi1FpyRzrxWnvmBfHZVxoozn0k6dCgnSrIigo2YFy6E5LdC9jNF9gqS75TEnkZ8YM3OVmOKmOsMiSLcq7EcCvlC1WcaCtAIHIO+uNKhkijaDbCnKZEZIUDGg37mK3lm5XccurxMiFh2VvHcorgm2kWQQ3BWIwJAgao9XUfpoBSp87V2030SR2CNoiwl6FwoZCNJBUTOlrLlKoqUoRCwtjGYvi/62JFzoUnmndZisROtxD96pN9NcyqWOUsTZgupcUq6qru723olGl+VxvrO/SfXYNT6ytRUleJdbWlsFjsxYBckQAmDsFYIDynF5toA2FVsBe9nfcNuuRZlPb0XAajSLA4X1oyRbu5NEVHkvQBRB3YbIQ4cizPeiEWKhByNmQAXZ1y77OuLnVoV4AX1W4IyOUNzpQ+OuJWo7JB10BSvKIY6V1nA4I4gGwpq6Ju6b4/6DrmRel+I6JdjDEaZRXm+WLuG8DTj9Ov2xy2yAM1gEu4EjQDgJCqo9uBGtazyEWmzmasspzHZJWEnbGG8OR0p4UHbA1I01y3I2d49seCh6l+nWl7ubl4Zj67PWWmP+G3yuntznOcYHZyS76w5rdkrUIabR1kZdQl/OIZY8RW5+DqEp9nToUOuUxK4JB/szPJocj5QV1mgKWU9lCH8dzjhMM6xvRfzFhpL6fEwngjARNKUrFuWwZAmzPN4t9mmf9LhPS2BKtmvTuF86zghnI5+K23aZbk/uvu0xoM8rRAE+XIf2sr850wKGUvSSTslmLbExEy+onPW8pFYYxVH1i1rdvgOQ5+vIIdS7FQU2hVN0QfOCbzzyQ08H6s/g3KiXcoNErvbvpUSv3q+XoguHeqkIA2pD9BR6DJVkh4VdSIzarue1yNX9nvf6vHsMO78XoIeUIkJGrhNnKEQnRI9F4So6bWjXqZ2iYsvlT6AJ3RB1qQXZ2Qk/ogfFen4R26qOcXwb25S47TVl/ivzYPPqiVMx7HJ28x982OPSYQ8mgkXQRrwwV7U3Ejlh14LQOkghjbcvxQRrFqIqJwgQuRwJNBF1WTIaRKmi1dbNR4QH496hvVEc8ji8yxdBd+cPqsuTIFemmEflvP8gh8gbHT8RH530DuRzpQLSRxzW/yqqv9oFoh6nn+unBFS6upmPeo8NmDvR1ZJ8tHtC4+oIsZ2+4f9+j9aedKgf6gGsNREao8fPJ3CIvvMIm3M5dKi6ng8jyLefILrjZziCNqfc7wuqkiPoLMqhBhnXwbbUVm/K2+klVrWKjWvVr1W1sFbVq1nDhgXjcCSdSA9BDo6ARbAEVsGFcBlcDdfAV+EBeByegXfgLxAqqjKpLFXOVK5SblIeUHYrTyrPKW+lMqnDUnL7oGeI6XG9DFc6/KW9qIO/HpUOPSMIkzZvmfkbN0mTGawrVHEydK7FsijtcC4zZ3xYE2+4M/P29GENR3nRIh7vf4siiSWx4jt0DCZUpnP6St909CQ3iHVxLplpotR2mcfJJlG0ee2yXMMTTJB8L4/xqBXkMeh02r618CrZTybb7Hnz5QYhk2jy/CIOiHo6sqDT1XePkC15RlSXiDyErgmeELLFAVOm97QiPpXNoK1OzzbPiU66+m3zMrwNk8vLFoyi+B+5W0rwAAB4nGNgZGBg4AFiMSBmYmAEwgVAzALmMQAACn4AzwAAAHicY2BgYGQAglv3rh8F0Y+EV9+F0QBdughQAAB4nGObz8DA7M7AwLgGiL8wMLApA/kvGBhYyhgYWKcwMDBcY2BgsoZgkDxIDpkGqYNgRlkEGzsG2wHEzJEQ/SCalRWCCellOoVQB9LL7gh06yaEPLIZyGyQf5DtAJmDTIPZ1oTdAHI3k/XfozA2yFxCboaFITIf5GfkMGBlZZTFhtkdsYszR6JikLtxqcWKp4DiCUKz+uLAWPQxWaNiou1Dx/GMssjpgIEBAAV1QmUAAA=="
